Add files via upload
Week 3 parts 4 and 5 completed. We now have the map covered in a shroud, which clears up as you explore. We have a FOV, so we can't see the entire dungeon. Enemies are now being created and placed on the map, and we have a basic turn based system in place, where the player takes a turn, followed by the enemies. We also have an attacking action, which will be put to use later on.